Move utilities from DynamoCore to DynamoUtilities.
