+++ src/DynamoCore/Core/NotificationObject.cs
