+++ src/DynamoCore/Core/Threading/Disposable.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Dynamo.Utilities
+namespace Dynamo.Core.Threading
 {
     public static class Disposable
     {
