+++ src/DynamoCore/DSEngine/CodeCompletionServices.cs
@@ -1,11 +1,13 @@
 ﻿using System.Diagnostics;
 using Dynamo.Utilities;
+using Dynamo.Nodes;
 using ProtoCore.DSDefinitions;
 using ProtoCore.Mirror;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using ProtoCore.Namespace;
+using Symbol = ProtoCore.Namespace.Symbol;
 
 namespace Dynamo.DSEngine.CodeCompletion
 {
