+++ src/DynamoUtilities/AssemblyConfiguration.cs
