+++ src/DynamoUtilities/DynamoUtilities.csproj
@@ -34,7 +34,12 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="ICSharpCode.AvalonEdit, Version=4.3.1.9429, Culture=neutral, PublicKeyToken=9cc39be672370310, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
+      <HintPath>..\..\extern\avalonEdit\ICSharpCode.AvalonEdit.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
+    <Reference Include="System.configuration" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
@@ -46,12 +51,24 @@
     <Compile Include="..\AssemblySharedInfoGenerator\AssemblySharedInfo.cs">
       <Link>Properties\AssemblySharedInfo.cs</Link>
     </Compile>
+    <Compile Include="AssemblyConfiguration.cs" />
     <Compile Include="AssemblyHelper.cs" />
     <Compile Include="DataMarshaler.cs" />
+    <Compile Include="EnumerableExtensions.cs" />
     <Compile Include="Extensions.cs" />
+    <Compile Include="Guid.cs" />
+    <Compile Include="ModifierKeys.cs" />
+    <Compile Include="Option.cs" />
+    <Compile Include="OrderedSet.cs" />
     <Compile Include="PathHelper.cs" />
+    <Compile Include="Point2D.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Rect2D.cs" />
+    <Compile Include="Thickness.cs" />
+    <Compile Include="TypeExtensions.cs" />
     <Compile Include="TypeSwitch.cs" />
+    <Compile Include="VersionUtilities.cs" />
+    <Compile Include="XmlElementHelper.cs" />
     <Compile Include="XmlHelper.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
