+++ src/DynamoUtilities/VersionUtilities.cs
@@ -1,10 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Autodesk.DesignScript.Interfaces;
-using Dynamo.DSEngine;
-using Dynamo.Models;
-using ProtoCore.Mirror;
 
 namespace Dynamo.Utilities
 {
