+++ src/DynamoUtilities/XmlElementHelper.cs
